requirements.txt
flask==3.0.2
selenium==4.30.0
pandas==2.2.1
python-dotenv==1.0.1
Flask-WTF==1.2.1
bootstrap-flask==2.3.3
